moved test data to a separate file to enable re-use by other modules
