+++ test/basic.js
@@ -1,155 +1,31 @@
 var test = require('tape')
-
+var data = require('./data')
 var osmDataToGeoJson = require('./osmdata-to-geojson')
+
 var collect = require('collect-stream')
 
 test('node', function (t) {
   t.plan(2)
 
-  var batch = [
-    {
-      type: 'node',
-      id: 1,
-      lat: 1.234,
-      lon: 4.321,
-      tags: {
-        interesting: 'this is'
-      }
-    }
-  ]
-
-  var expected = {
-    type: 'FeatureCollection',
-    features: [
-      {
-        type: 'Feature',
-        properties: {
-          interesting: 'this is'
-        },
-        geometry: {
-          type: 'Point',
-          coordinates: [4.321, 1.234]
-        }
-      }
-    ]
-  }
-
-  osmDataToGeoJson(batch, function (err, geojson) {
+  osmDataToGeoJson(data.node.batch, function (err, geojson) {
     t.error(err)
-    t.deepEqual(geojson, expected)
+    t.deepEqual(geojson, data.node.expected)
   })
 })
 
 test('way', function (t) {
   t.plan(2)
 
-  var batch = [
-    {
-      type: 'way',
-      id: 'A',
-      tags: {
-        interesting: 'this is'
-      },
-      nodes: ['B', 'C', 'D']
-    },
-    {
-      type: 'node',
-      id: 'B',
-      lat: 0.0,
-      lon: 1.0
-    },
-    {
-      type: 'node',
-      id: 'C',
-      lat: 0.0,
-      lon: 1.1
-    },
-    {
-      type: 'node',
-      id: 'D',
-      lat: 0.1,
-      lon: 1.2
-    }
-  ]
-
-  var expected = {
-    type: 'FeatureCollection',
-    features: [
-      {
-        type: 'Feature',
-        properties: {
-          interesting: 'this is'
-        },
-        geometry: {
-          type: 'LineString',
-          coordinates: [
-            [1.0, 0.0],
-            [1.1, 0.0],
-            [1.2, 0.1]
-          ]
-        }
-      }
-    ]
-  }
-
-  osmDataToGeoJson(batch, function (err, geojson) {
+  osmDataToGeoJson(data.way.batch, function (err, geojson) {
     t.error(err)
-    t.deepEqual(geojson, expected)
+    t.deepEqual(geojson, data.way.expected)
   })
 })
 
-
 test('way (streaming)', function (t) {
   t.plan(4)
-
-  var batch = [
-    {
-      type: 'way',
-      id: 'A',
-      tags: {
-        interesting: 'this is'
-      },
-      nodes: ['B', 'C', 'D']
-    },
-    {
-      type: 'node',
-      id: 'B',
-      lat: 0.0,
-      lon: 1.0
-    },
-    {
-      type: 'node',
-      id: 'C',
-      lat: 0.0,
-      lon: 1.1
-    },
-    {
-      type: 'node',
-      id: 'D',
-      lat: 0.1,
-      lon: 1.2
-    }
-  ]
-
-  var expected = {
-    type: 'FeatureCollection',
-    features: [
-      {
-        type: 'Feature',
-        properties: {
-          interesting: 'this is'
-        },
-        geometry: {
-          type: 'LineString',
-          coordinates: [
-            [1.0, 0.0],
-            [1.1, 0.0],
-            [1.2, 0.1]
-          ]
-        }
-      }
-    ]
-  }
+  var batch = data.way.batch
+  var expected = data.way.expected
 
   testStreaming()
 
@@ -181,97 +57,14 @@ test('way (streaming)', function (t) {
 test('polygon', function (t) {
   t.plan(2)
 
-  var batch = [
-    {
-      type: 'way',
-      id: 'A',
-      nodes: ['B', 'C', 'D', 'E', 'B'],
-      tags: {area: 'yes'}
-    },
-    {
-      type: 'node',
-      id: 'B',
-      lat: 0.0,
-      lon: 0.0
-    },
-    {
-      type: 'node',
-      id: 'C',
-      lat: 0.0,
-      lon: 1.0
-    },
-    {
-      type: 'node',
-      id: 'D',
-      lat: 1.0,
-      lon: 1.0
-    },
-    {
-      type: 'node',
-      id: 'E',
-      lat: 1.0,
-      lon: 0.0
-    }
-  ]
-
-  var expected = {
-    type: 'FeatureCollection',
-    features: [
-      {
-        type: 'Feature',
-        properties: {
-          area: 'yes'
-        },
-        geometry: {
-          type: 'Polygon',
-          coordinates: [[
-            [0.0, 0.0],
-            [1.0, 0.0],
-            [1.0, 1.0],
-            [0.0, 1.0],
-            [0.0, 0.0]
-          ]]
-        }
-      }
-    ]
-  }
-
-  osmDataToGeoJson(batch, function (err, geojson) {
+  osmDataToGeoJson(data.polygon.batch, function (err, geojson) {
     t.error(err)
-    t.deepEqual(geojson, expected)
+    t.deepEqual(geojson, data.polygon.expected)
   })
 })
 
 test('opts.map', function (t) {
   t.plan(2)
-
-  var batch = [
-    {
-      type: 'node',
-      id: 1,
-      lat: 1.234,
-      lon: 4.321,
-      tags: {
-        interesting: 'this is'
-      }
-    }
-  ]
-
-  var expected = {
-    type: 'FeatureCollection',
-    features: [
-      {
-        type: 'Feature',
-        properties: {
-          interesting: 'this is'
-        },
-        geometry: {
-          type: 'Point',
-          coordinates: [4.321, 1.234]
-        }
-      }
-    ]
-  }
 
   function mapFn (geom) {
     if (geom.id) {
@@ -280,9 +73,9 @@ test('opts.map', function (t) {
     return geom
   }
 
-  osmDataToGeoJson(batch, { map: mapFn }, function (err, geojson) {
+  osmDataToGeoJson(data.node.batch, { map: mapFn }, function (err, geojson) {
     t.error(err)
-    t.deepEqual(geojson, expected)
+    t.deepEqual(geojson, data.node.expected)
   })
 })
 
@@ -328,7 +121,7 @@ test('invalid polygon', function (t) {
     ]
   }
 
-  osmDataToGeoJson(batch, function (err, geojson) {
+  osmDataToGeoJson(data.json2batch(batch), function (err, geojson) {
     t.error(err)
     t.deepEqual(geojson, expected)
   })
